+++ nodes/Miva/Miva.node.ts
@@ -4,7 +4,7 @@ import type {
 	INodeType,
 	INodeTypeDescription,
 } from 'n8n-workflow';
-import { NodeConnectionType, NodeOperationError } from 'n8n-workflow';
+import { NodeOperationError } from 'n8n-workflow';
 import { operations } from './operations';
 import {
 	OPERATION_PARAMETER,
@@ -65,8 +65,8 @@ export class Miva implements INodeType {
 		defaults: {
 			name: 'Miva',
 		},
-		inputs: [NodeConnectionType.Main],
-		outputs: [NodeConnectionType.Main],
+		inputs: ['main'],
+		outputs: ['main'],
 		credentials: [
 			{
 				name: 'mivaApi',